% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_rows_not_all_equal}
\alias{filter_rows_not_all_equal}
\title{Title}
\usage{
filter_rows_not_all_equal(
  input_tibble,
  columns_to_check = NULL,
  columns_to_exclude = NULL
)
}
\arguments{
\item{columns_to_check}{}
}
\description{
Title
}
\examples{
# Sample data
df <- tibble(
  ID = c(1, 2, 3, 4, 5, 6),
  A = c(10, 20, 10, 30, 5, 6),
  B = c(5, 5, 5, 5, 5, 5),
  C = c("X", "Y", "X", "Z", "Z", "Z")
)

#Only check columns A and B
result_df1 <- filter_rows_not_all_equal(df, columns_to_check = c("A", "B"))
print(result_df1)

# Check all columns except ID
result_df2 <- filter_rows_not_all_equal(df, columns_to_exclude = c("C", "ID"))
print(result_df2)

# Check all columns
result_df3 <- filter_rows_not_all_equal(df)
print(result_df3)
}
