% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_equal_column}
\alias{find_equal_column}
\title{Find column name equal to vector, if ant}
\usage{
find_equal_column(input_tibble, target_vector)
}
\arguments{
\item{target_vector}{}
}
\value{
A column name, or character(0) if there is no match
}
\description{
Find column name equal to vector, if ant
}
\examples{
# Sample data
df <- tibble(
  A = c(1, 2, 1),
  B = c(4, 5, 6),
  C = c(7, 8, 9)
)

# Target vector to compare
target_vector <- c(1, 2, 1)

# Find columns equal to the target vector
matching_columns <- find_equal_column(df, target_vector)

# Print the result
print(matching_columns)
}
