% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_diagnostics.R
\name{gamma_trace}
\alias{gamma_trace}
\title{Title}
\usage{
gamma_trace(
  number_of_variants = 1,
  nchains = 5,
  include_warmup = FALSE,
  prefix = "",
  variants = seq_len(number_of_variants),
  samples = c(1)
)
}
\arguments{
\item{number_of_variants}{The number of variants considered.}

\item{nchains}{Number of chains to display (the code will display chains 1:nchains)}

\item{include_warmup}{Whether to display the warm up iterations as well}

\item{prefix}{The path to the folder results}

\item{variants}{A integer vector describing which variants to plot. By default, all variants.}

\item{samples}{A integer vector describing which samples to plot. By default, only the first sample.}
}
\description{
Title
}
