% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_diagnostics.R
\name{gamma_trace_with_variants_identified}
\alias{gamma_trace_with_variants_identified}
\title{Trace plot for the relative abundances for all chains, with chain-specific variants disambiguated.}
\usage{
gamma_trace_with_variants_identified(
  number_of_variants = 1,
  nchains = 5,
  include_warmup = FALSE,
  prefix = "",
  variants = NULL,
  samples = c(1)
)
}
\arguments{
\item{number_of_variants}{The number of variants considered.}

\item{nchains}{Number of chains to display (the code will display chains 1:nchains)}

\item{include_warmup}{Whether to display the warm up iterations as well}

\item{prefix}{The path to the folder results}

\item{variants}{A integer vector describing which variants to plot. By default, all variants.}

\item{samples}{A integer vector describing which samples to plot. By default, only the first sample.}
}
\value{
A trace plot for the relative abundances
}
\description{
Trace plot for the relative abundances for all chains, with chain-specific variants disambiguated.
}
