% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_diagnostics.R
\name{log_likelihood_trace}
\alias{log_likelihood_trace}
\title{Traceplot for the log likelihood}
\usage{
log_likelihood_trace(
  number_of_variants = 1,
  nchains = 5,
  include_warmup = FALSE,
  prefix = ""
)
}
\arguments{
\item{number_of_variants}{The number of variants considered.}

\item{nchains}{Number of chains to display (the code will display chains 1:nchains)}

\item{include_warmup}{Whether to display the warm up iterations as well}

\item{prefix}{The path to the folder results}
}
\value{
A trace plot for the log likelihood
}
\description{
Traceplot for the log likelihood
}
