% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_diagnostics.R
\name{nucleotide_count_trace}
\alias{nucleotide_count_trace}
\title{Trace plot for the count of each type of nucleotide}
\usage{
nucleotide_count_trace(
  number_of_variants = 1,
  nchains = 5,
  include_warmup = FALSE,
  prefix = ""
)
}
\arguments{
\item{number_of_variants}{The number of variants considered.}

\item{nchains}{Number of chains to display (the code will display chains 1:nchains)}

\item{include_warmup}{Whether to display the warm up iterations as well}

\item{prefix}{The path to the folder results}
}
\value{
A trace plot for the count of each type of nucleotide
}
\description{
Trace plot for the count of each type of nucleotide
}
