% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_visualisation.R
\name{plot_relative_abundance}
\alias{plot_relative_abundance}
\title{Plot the relative abundance for each variants in each sample}
\usage{
plot_relative_abundance(number_of_variants = 1, chain = 5, prefix = "")
}
\arguments{
\item{number_of_variants}{The number of variants considered, this is used to choose the file path}

\item{chain}{The chain selected. This ought to be the chain with the lowest deviance among all chains.}

\item{prefix}{File path describing where the results are stored.}
}
\value{
A plot the relative abundance for each variants in each sample
}
\description{
Plot the relative abundance for each variants in each sample
}
